requirements.txt
APScheduler==3.5.1
BareNecessities==0.2.8
Flask==0.12.2
Flask-Cors==3.0.3
Flask-Mail==0.9.1
Jinja2==2.10
Mail==2.1.0
MarkupSafe==1.0
PyJWT==1.6.1
PyNaCl==1.2.1
Werkzeug==0.14.1
app==0.0.1
blinker==1.4
cffi==1.14.0
click==6.7
coverage==4.5.1
idna==2.6
itsdangerous==0.24
pip==9.0.1
pyasn1==0.4.2
pycparser==2.18
pytz==2018.4
setuptools==40.2.0
six==1.11.0
tzlocal==2.0.0b1
validate-email==1.3
Flask-PyMongo==2.2.0
numpy==1.15.4